requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
websockets>=12.0
google-genai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.28.0
